feat: Add getPcuDateResetDrugStock function to retrieve date_reset_drug_stock from directus
This commit adds a new function getPcuDateResetDrugStock to the dateResetDrug controller. This function retrieves the date_reset_drug_stock value from the directus database for a given pcucode. If the pcucode is not provided, the function returns undefined.

The function makes use of the readItem function from the @tspvivek/refine-directus package and the directusClient instance from the ../directusClient module.

This change is necessary to fetch the date_reset_drug_stock value for further processing in the application.

Ref: #123